refactor: move data-quality profiling out of utils into its own module
